chart_scheduler/requirements.txt
uvicorn[standard]
gunicorn
fastapi
finance-datareader
bs4
sqlalchemy
PyJWT
bcrypt
databases
aiomysql
pykrx
matplotlib
rocketry
asyncio